Changed DT-CRDT to use faster btree impl. Changed run-on-old to merge delete operations. Added stats tracking. Perf increased. Disabled positional updates & OT code in DT-CRDT - see dt-crdt-fully-working branch for older version of the code.

+++ crates/diamond-types-crdt/src/crdtspan.rs
@@ -1,7 +1,7 @@
+use content_tree::ContentLength;
 use diamond_core_old::CRDTId;
 use rle::{HasLength, MergableSpan, SplitableSpanHelpers};
 
-use content_tree::ContentLength;
 use rle::Searchable;
 
 #[derive(Debug, Copy, Clone, Default, Eq, PartialEq)]
@@ -20,7 +20,7 @@ impl Searchable for CRDTSpan {
         if self.loc.agent == loc.agent
             && loc.seq >= self.loc.seq
             && loc.seq < self.loc.seq + self.len {
-            Some((loc.seq - self.loc.seq) as usize)
+            Some(loc.seq - self.loc.seq)
         } else { None }
     }
